requirements.txt
 streamlit 
 langchain 
 python-pptx 
 requests 
 python-dotenv
 groq